requirements.txt
customtkinter>=5.2.1
python-snap7>=1.0.0
matplotlib>=3.8.0
numpy>=1.26.0
mysql-connector-python